Make linter happy and remove unnecessary scripts
This commit makes our shiny new ES6 linter
(using the airbnb config - see previous commit)
happy. The linter's also set to be really mean -
if it's not happy, the web-app won't build.

build.js was removed as we don't currently need it.

The most signficant formatting change was the renaming
of JS files containing JSX to .JSX from .JS, which is
good practice.
