+++ .eslintrc.json
@@ -1,3 +1,7 @@
 {
-    "extends": "airbnb"
+    "extends": "airbnb",
+    "env": {
+        "es6": true,
+        "browser": true
+    }
 }
