+++ config/polyfills.js
@@ -1,11 +1,12 @@
-'use strict';
+const rejectionTracking = require('promise/lib/rejection-tracking');
+const es6Extensions = require('promise/lib/es6-extensions.js');
 
 if (typeof Promise === 'undefined') {
   // Rejection tracking prevents a common issue where React gets into an
   // inconsistent state due to an error, but it gets swallowed by a Promise,
   // and the user has no idea what causes React's erratic future behavior.
-  require('promise/lib/rejection-tracking').enable();
-  window.Promise = require('promise/lib/es6-extensions.js');
+  rejectionTracking.enable();
+  window.Promise = es6Extensions;
 }
 
 // fetch() polyfill for making API calls.
